dse_predict/requirements.txt
Flask==0.12.2
google-cloud-datastore==1.4.0
google-cloud-storage==1.6.0
gunicorn==19.7.1
oauth2client==4.1.2
six==1.11.0
numpy==1.14.0
scipy>=0.15.1
scikit-learn>=0.19.1
